+++ .prettierignore
@@ -29,3 +29,7 @@ CHANGELOG.md
 # temp volumes
 .hive/
 .hive-dev/
+
+
+# File generated by @hive/app#generate-changelog script
+packages/web/app/src/components/ui/changelog/generated-changelog.ts
