+++ packages/web/app/package.json
@@ -4,13 +4,13 @@
   "private": true,
   "scripts": {
     "analyze": "pnpm build:config && ANALYZE=1 next build",
-    "build": "pnpm generate-changelogs && pnpm build:config && BUILD=1 next build && tsx ../../../scripts/runify.ts",
+    "build": "pnpm generate-changelog && pnpm build:config && BUILD=1 next build && tsx ../../../scripts/runify.ts",
     "build-storybook": "storybook build",
     "build:config": "tsup-node --no-splitting --out-dir . --loader \".mts=ts\" --format esm --target node21 next.config.mts",
-    "dev": "pnpm generate-changelogs && pnpm build:config && next dev | pino-pretty",
-    "generate-changelogs": "node ../../../scripts/generate-changelogs.js && pnpm prettier",
+    "dev": "pnpm generate-changelog && pnpm build:config && next dev | pino-pretty",
+    "generate-changelog": "node ../../../scripts/generate-changelog.js",
     "postbuild": "pnpm --filter @hive/app deploy --prod --no-optional deploy-tmp/web && rimraf dist/node_modules && mv deploy-tmp/web/node_modules dist && rimraf deploy-tmp/web",
-    "prepare": "pnpm run generate-changelogs",
+    "postinstall": "pnpm generate-changelog",
     "start": "node dist/index.js",
     "storybook": "storybook dev -p 6006",
     "typecheck": "tsc"
