+++ packages/web/app/public/__ENV.js
@@ -1,0 +1,1 @@
+window.__ENV = {"NODE_ENV":"production","GRAPHQL_PUBLIC_ENDPOINT":"http://localhost:3001/graphql","APP_BASE_URL":"http://localhost:3000","ENVIRONMENT":"development","AUTH_GITHUB":"","AUTH_GOOGLE":"","AUTH_OKTA":"0","AUTH_OKTA_HIDDEN":"0","AUTH_ORGANIZATION_OIDC":"0","INTEGRATION_SLACK":"0","STRIPE_PUBLIC_KEY":"","MEMBER_ROLES_DEADLINE":"2024-01-31","GRAPHQL_PUBLIC_ORIGIN":"http://localhost:3001"};
