+++ packages/web/app/src/components/ui/user-menu.tsx
@@ -35,7 +35,7 @@ import { GetStartedProgress } from '../get-started/wizard';
 import { MemberRoleMigrationStickyNote } from '../organization/members/migration';
 import { UserSettingsModal } from '../user/settings';
 import { Changelog } from './changelog/changelog';
-import { latestChangelogs } from './changelog/generated-changelogs';
+import { latestChangelog } from './changelog/generated-changelog';
 
 export const UserMenu_CurrentOrganizationFragment = graphql(`
   fragment UserMenu_CurrentOrganizationFragment on Organization {
@@ -113,7 +113,7 @@ export function UserMenu(props: {
         />
       ) : null}
       <div className="flex flex-row items-center gap-8">
-        <Changelog changes={latestChangelogs} />
+        <Changelog changes={latestChangelog} />
         <MemberRoleMigrationStickyNote organization={currentOrganization} />
         {currentOrganization ? <GetStartedProgress tasks={currentOrganization.getStarted} /> : null}
         <DropdownMenu>
