+++ packages/web/app/src/stories/changelog.stories.tsx
@@ -12,7 +12,6 @@ type Story = StoryObj<typeof Changelog>;
 const changes = [
   {
     date: 'February 26, 2024',
-    userReadItem: false,
     href: 'https://the-guild.dev/graphql/hive/product-updates/2024-02-26-schema-check-top-affected-operations',
     title: 'Top affected operations based on conditional breaking change configuration',
     description:
@@ -20,7 +19,6 @@ const changes = [
   },
   {
     date: 'February 8, 2024',
-    userReadItem: false,
     href: 'https://the-guild.dev/graphql/hive/product-updates/2024-02-08-upcoming-stitching-project-changes',
     title: 'Upcoming changes to schema composition for Schema Stitching projects',
     description:
